TGZ Game Controller Support
Added TGZ Game Controller to manually control the turret which now has a camera attached. Way more fun with the cat.

Need to revisit the Lidar control and FIFO reading. The previous firing issue was worked around by flashing a Standard Firmata to an Arduino Uno Rev3 to handle the output to SSR and thus turning the solenoid off or on.